Drop lower-constraints.txt and it's testing
As discussed in TC PTG[1] and TC resolution[2], dropping
the lower-constraints.txt file and its testing.
We will keep lower bounds in the requirements.txt file but
with a note that these are not tested lower bounds and we
try our best to keep them updated.

And switch zuul task openstack-python3-zed-jobs

[1] https://etherpad.opendev.org/p/tc-zed-ptg#L326
[2] https://governance.openstack.org/tc/resolutions/20220414-drop-lower-constraints.html#proposal

Change-Id: I6f467a6bb175e27c011a57fec15a37b7c8dabdd0